+++ Packages/ai.natml.vision.yunet/Changelog.md
@@ -1,3 +1,8 @@
+## 1.0.3
++ Added `YuNetPredictor.Create` static method for creating the YuNet predictor.
++ Removed `YuNetPredictor` public constructor. Use the `Create` static method instead.
++ Upgraded to NatML 1.1.15.
+
 ## 1.0.2
 + Upgraded to NatML 1.0.11.